Reorganized the UserController => update functions, putting them in their own helper file. Also added the Tracker model and a one-way-many association TwitterAccount -> Tracker
