+++ views/index.ejs
@@ -18,16 +18,16 @@
           <a id="page-top"></a>
         </li>
         <li>
-          <a ng-click="scrollTo('content')">Content</a>
+          <a href="#" ng-click="scrollTo('content')">Content</a>
         </li>
         <li>
-          <a ng-click="scrollTo('audience')">Audience</a>
+          <a href="#" ng-click="scrollTo('audience')">Audience</a>
         </li>
         <li>
-          <a ng-click="scrollTo('trends')">Trends</a>
+          <a href="#" ng-click="scrollTo('trends')">Trends</a>
         </li>
         <li>
-          <a ng-click="scrollTo('cta')">CTA</a>
+          <a href="#" ng-click="scrollTo('cta')">CTA</a>
         </li>
       </ul>
       <ul ng-cloak ng-if="!$root.currentUser" class="nav navbar-nav navbar-right">
